cloud_run_jobs/moi_codebase_population/requirements.txt

pandas==1.5.2
requests==2.28.1
google-cloud-bigquery==3.4.0
